+++ lib/core/validator.dart
@@ -1,6 +1,6 @@
-// ignore_for_file: public_member_api_docs, sort_constructors_first
+part of 'package:flutter_reactive/flutter_reactive.dart';
 
-class ValidatorError<T> implements Exception {
+class ReactiveValidatorError<T> implements Exception {
   final String? _message;
   final T? value;
 
@@ -9,21 +9,24 @@ class ValidatorError<T> implements Exception {
           ? "ValidatorError: INVALID_VALUE ${value != null ? "($value)" : ""}"
           : _message!;
 
-  ValidatorError({String? message, this.value}) : _message = message;
+  ReactiveValidatorError({String? message, this.value}) : _message = message;
 
   @override
   String toString() => message;
 }
 
-class Validator<T> {
+/// Validator for Reactive values, allowing for custom validation logic and error handling.
+class ReactiveValidator<T> {
   final bool Function(T v) validator;
   final String? message;
 
-  Validator(this.validator, [this.message]);
+  ReactiveValidator(this.validator, [this.message]);
 
   bool run(T v) {
     if (validator(v)) return true;
-    if (message != null) throw ValidatorError(value: v, message: message);
+    if (message != null) {
+      throw ReactiveValidatorError(value: v, message: message);
+    }
     return false;
   }
 }